+++ src/eye.js
@@ -393,6 +393,17 @@ function EyeElement() {
    * @type {Map<String,Set<{callback: function, target: string}>>}
    */
   this.dlgListeners = new Map();
+
+  let normalSetterGetter = (action, v, elm) => v;
+  /**
+   * Custom way or modifier that redefine the way you set/get
+   * this element `textContent` or `value`:
+   * - access this feature from `.redefine` method.
+   */
+  this.customSet = {
+    value: normalSetterGetter,
+    text: normalSetterGetter
+  };
 }
 EyeElement.prototype = {
   /**
@@ -520,8 +531,8 @@ EyeElement.prototype = {
   text: function (text) {
     let out = "";
     (this.raw instanceof NodeList ? [...this.raw.entries()] : [[0, this.raw]]).forEach(([idx, elm]) => {
-      if (!text) return out = elm.textContent;
-      elm.textContent = text;
+      if (!text) return out = this.customSet.text("get", elm.textContent, elm);
+      elm.textContent = this.customSet.text("set", text, elm);
     })
     return out ? out : this;
   },
@@ -926,8 +937,9 @@ EyeElement.prototype = {
    * @returns {EyeElement|null}
    */
   child: function (index) {
-    let it = (this.raw instanceof NodeList ? this.raw.item(0) : this.raw).children[index];
-    if (it) return eye(it);
+    let it = (this.raw instanceof NodeList ? this.raw.item(0) : this.raw);
+    if (index === undefined) return it.children.length;
+    if (it.children[index]) return eye(it.children[index]);
     return null;
   },
   /**
@@ -937,8 +949,11 @@ EyeElement.prototype = {
    * @returns 
    */
   val: function (value) {
-    if (value) (this.raw instanceof NodeList ? [...this.raw.entries()] : [[0, this.raw]]).forEach(([idx, a]) => a.value = value);
-    else return (this.raw instanceof NodeList ? this.raw.item(0) : this.raw).value;
+    if (value) (this.raw instanceof NodeList ? [...this.raw.entries()] : [[0, this.raw]]).forEach(([idx, a]) => a.value = this.customSet.value("set", value, a));
+    else {
+      let it = (this.raw instanceof NodeList ? this.raw.item(0) : this.raw);
+      return this.customSet.value("get", it.value, it);
+    }
     return this;
   },
   /**
@@ -977,6 +992,19 @@ EyeElement.prototype = {
       out.url = new URLSearchParams(out.json).toString();
       return out;
     } else console.warn(`[EyeJS] this is a multi selection, it's not serializable!`);
+  },
+  /**
+   * Redefine the way `.text` or `.val` set or get data to and from this element.
+   * @method EyeElement#redefine
+   * @param {"text" | "value"} type 
+   * @param {(action: "set" | "get", value: *, elm: EyeElement) => *} process 
+   */
+  redefine: function (type, process) {
+    if (["text", "value"].includes(type) && typeof process == "function")
+      this.customSet[type] = process;
+
+    
+    return this;
   }
 };